Fix HeroSection
